requirements.txt
streamlit
google-generativeai
python-dotenv
langchain
pyPDF2
chromadb